Changed header styling, added missing typing, reorganized alert file structure
